fix(server): resolve release build errors; 404 for unknown routes
The v2.26.0/v2.26.1 tags were cut while these fixes only existed in the
working tree, so every release build failed with 14 compile errors and
ghcr latest stayed on the v2.17.15 binary (which panics on startup under
axum 0.8's route-syntax check).

Also scope auth/rate-limit middleware with route_layer so unmatched
paths return 404 instead of 401 (fixes nonexistent_route_returns_404).

+++ server/crates/muorg-server/src/transcode.rs
@@ -52,7 +52,7 @@ fn do_transcode(
         .as_ref()
         .and_then(|p| p.audio())
         .ok_or("No audio codec parameters")?;
-    let sample_rate = config.sample_rate as u32;
+    let sample_rate = config.sample_rate;
     let channels = audio_params
         .channels
         .clone()
